requirements/base.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==5.0.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.34
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.25.0
